Move `get_mid_point` to utils module
